move functions, delete files, create hierarchy
