--- policy_ddpg.py
@@ -1,69 +1,0 @@
-"""
-Example of running PyTorch implementation of DDPG on HalfCheetah.
-"""
-from gym.envs.mujoco import HalfCheetahEnv
-
-from rlkit.envs.wrappers import NormalizedBoxEnv
-from quad_env import QuadEnv
-from rlkit.exploration_strategies.base import (
-    PolicyWrappedWithExplorationStrategy
-)
-from rlkit.exploration_strategies.ou_strategy import OUStrategy
-from rlkit.launchers.launcher_util import setup_logger
-from rlkit.torch.networks import FlattenMlp, TanhMlpPolicy
-from rlkit.torch.ddpg.ddpg import DDPG
-import rlkit.torch.pytorch_util as ptu
-
-
-def experiment(variant):
-    env = QuadEnv()
-    env = NormalizedBoxEnv(env)
-    # Or for a specific version:
-    # import gym
-    # env = NormalizedBoxEnv(gym.make('HalfCheetah-v1'))
-    es = OUStrategy(action_space=env.action_space)
-    obs_dim = env.observation_space.low.size
-    action_dim = env.action_space.low.size
-    qf = FlattenMlp(
-        input_size=obs_dim + action_dim,
-        output_size=1,
-        hidden_sizes=[400, 300],
-    )
-    policy = TanhMlpPolicy(
-        input_size=obs_dim,
-        output_size=action_dim,
-        hidden_sizes=[400, 300],
-    )
-    exploration_policy = PolicyWrappedWithExplorationStrategy(
-        exploration_strategy=es,
-        policy=policy,
-    )
-    algorithm = DDPG(
-        env,
-        qf=qf,
-        policy=policy,
-        exploration_policy=exploration_policy,
-        **variant['algo_params']
-    )
-    algorithm.to(ptu.device)
-    algorithm.train()
-
-
-if __name__ == "__main__":
-    # noinspection PyTypeChecker
-    variant = dict(
-        algo_params=dict(
-            num_epochs=200,
-            num_steps_per_epoch=500,
-            num_steps_per_eval=1000,
-            use_soft_update=True,
-            tau=1e-2,
-            batch_size=128,
-            max_path_length=100,
-            discount=0.99,
-            qf_learning_rate=1e-3,
-            policy_learning_rate=1e-4,
-        ),
-    )
-    setup_logger('ddpg_quad', variant=variant)
-    experiment(variant)
